Refactor pkg/crane/api -> pkg/crane
