+++ pkg/crane/copy.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package api
+package crane
 
 import (
 	"github.com/google/go-containerregistry/pkg/authn"
